test: v1_keys_getVerifications
