+++ apps/api/src/pkg/testutil/test_route_roles.ts
@@ -38,6 +38,7 @@ export function runSharedRoleTests<TReq>(config: {
       h.useRoutes(config.registerHandler);
 
       const req = await config.prepareRequest(h);
+      console.log(req);
       const res = await h.do<TReq, ErrorResponse>(req);
       console.log(res);
       expect(res.status).toEqual(403);
