+++ apps/api/src/routes/v1_keys_getVerifications.ts
@@ -4,6 +4,7 @@ import { createRoute, z } from "@hono/zod-openapi";
 
 import { rootKeyAuth } from "@/pkg/auth/root_key";
 import { UnkeyApiError, openApiErrorResponses } from "@/pkg/errors";
+import { buildQuery } from "@unkey/rbac";
 
 const route = createRoute({
   method: "get",
@@ -74,9 +75,6 @@ export type V1KeysGetVerificationsResponse = z.infer<
 >;
 export const registerV1KeysGetVerifications = (app: App) =>
   app.openapi(route, async (c) => {
-    const auth = await rootKeyAuth(c);
-
-    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
     const { keyId, ownerId, start, end } = c.req.query();
 
     const ids: {
@@ -150,6 +148,15 @@ export const registerV1KeysGetVerifications = (app: App) =>
 
       ids.push(...keys.map(({ key, api }) => ({ keyId: key.id, apiId: api.id })));
     }
+
+    const apiIds = Array.from(new Set(ids.map(({ apiId }) => apiId)));
+    const auth = await rootKeyAuth(
+      c,
+      buildQuery(({ or, and }) =>
+        or("*", "api.*.read_key", and(...apiIds.map((apiId) => `api.${apiId}.read_key`))),
+      ),
+    );
+    const authorizedWorkspaceId = auth.authorizedWorkspaceId;
 
     const verificationsFromAllKeys = await Promise.all(
       ids.map(({ keyId, apiId }) => {
